# trie: refactor stacktrie