--- a/trie/stacktrie_marshalling.go
+++ b/trie/stacktrie_marshalling.go
@@ -23,60 +23,66 @@ import (
 	"encoding/gob"
 )
 
-var ( //Compile-time interface checks
+// Compile-time interface checks.
+var (
 	_ = encoding.BinaryMarshaler((*StackTrie)(nil))
 	_ = encoding.BinaryUnmarshaler((*StackTrie)(nil))
 )
 
 // NewFromBinaryV2 initialises a serialized stacktrie with the given db.
 // OBS! Format was changed along with the name of this constructor.
-func NewFromBinaryV2(data []byte, writeFn NodeWriteFunc) (*StackTrie, error) {
-	stack := NewStackTrie(writeFn)
+func NewFromBinaryV2(data []byte) (*StackTrie, error) {
+	stack := NewStackTrie(nil)
 	if err := stack.UnmarshalBinary(data); err != nil {
 		return nil, err
 	}
 	return stack, nil
 }
 
-// UnmarshalBinary implements encoding.BinaryMarshaler
-func (st *StackTrie) MarshalBinary() (data []byte, err error) {
+// MarshalBinary implements encoding.BinaryMarshaler.
+func (t *StackTrie) MarshalBinary() (data []byte, err error) {
 	var (
 		b bytes.Buffer
 		w = bufio.NewWriter(&b)
 	)
-	if err := gob.NewEncoder(w).Encode(st.owner); err != nil {
+	if err := gob.NewEncoder(w).Encode(t.owner); err != nil {
 		return nil, err
 	}
-	if err := st.root.marshalInto(w); err != nil {
+	if err := t.root.marshalInto(w); err != nil {
 		return nil, err
 	}
 	w.Flush()
 	return b.Bytes(), nil
 }
 
-// UnmarshalBinary implements encoding.BinaryUnmarshaler
-func (stack *StackTrie) UnmarshalBinary(data []byte) error {
+// UnmarshalBinary implements encoding.BinaryUnmarshaler.
+func (t *StackTrie) UnmarshalBinary(data []byte) error {
 	r := bytes.NewReader(data)
-	if err := gob.NewDecoder(r).Decode(&stack.owner); err != nil {
+	if err := gob.NewDecoder(r).Decode(&t.owner); err != nil {
 		return err
 	}
-	if err := stack.root.unmarshalFrom(r); err != nil {
+	if err := t.root.unmarshalFrom(r); err != nil {
 		return err
 	}
 	return nil
 }
 
-type encodedNode struct {
-	NodeType uint8
-	Val      []byte
-	Key      []byte
+type stackNodeMarshaling struct {
+	Typ uint8
+	Key []byte
+	Val []byte
 }
 
-func (st *stNode) marshalInto(w *bufio.Writer) (err error) {
-	if err := gob.NewEncoder(w).Encode(encodedNode{st.nodeType, st.val, st.key}); err != nil {
+func (n *stNode) marshalInto(w *bufio.Writer) (err error) {
+	enc := stackNodeMarshaling{
+		Typ: n.typ,
+		Key: n.key,
+		Val: n.val,
+	}
+	if err := gob.NewEncoder(w).Encode(enc); err != nil {
 		return err
 	}
-	for _, child := range st.children {
+	for _, child := range n.children {
 		if child == nil {
 			w.WriteByte(0)
 			continue
@@ -89,16 +95,16 @@ func (st *stNode) marshalInto(w *bufio.Writer) (err error) {
 	return nil
 }
 
-func (st *stNode) unmarshalFrom(r *bytes.Reader) error {
-	var dec encodedNode
+func (n *stNode) unmarshalFrom(r *bytes.Reader) error {
+	var dec stackNodeMarshaling
 	if err := gob.NewDecoder(r).Decode(&dec); err != nil {
 		return err
 	}
-	st.nodeType = dec.NodeType
-	st.val = dec.Val
-	st.key = dec.Key
+	n.typ = dec.Typ
+	n.key = dec.Key
+	n.val = dec.Val
 
-	for i := range st.children {
+	for i := range n.children {
 		if b, err := r.ReadByte(); err != nil {
 			return err
 		} else if b == 0 {
@@ -108,7 +114,7 @@ func (st *stNode) unmarshalFrom(r *bytes.Reader) error {
 		if err := child.unmarshalFrom(r); err != nil {
 			return err
 		}
-		st.children[i] = &child
+		n.children[i] = &child
 	}
 	return nil
 }