--- a/trie/stacktrie_test.go
+++ b/trie/stacktrie_test.go
@@ -379,45 +379,48 @@ func TestStacktrieNotModifyValues(t *testing.T) {
 
 // TestStacktrieSerialization tests that the stacktrie works well if we
 // serialize/unserialize it a lot
-//func TestStacktrieSerialization(t *testing.T) {
-//	var (
-//		st       = NewStackTrie(nil)
-//		nt       = NewEmpty(NewDatabase(rawdb.NewMemoryDatabase(), nil))
-//		keyB     = big.NewInt(1)
-//		keyDelta = big.NewInt(1)
-//		vals     [][]byte
-//		keys     [][]byte
-//	)
-//	getValue := func(i int) []byte {
-//		if i%2 == 0 { // large
-//			return crypto.Keccak256(big.NewInt(int64(i)).Bytes())
-//		} else { //small
-//			return big.NewInt(int64(i)).Bytes()
-//		}
-//	}
-//	for i := 0; i < 10; i++ {
-//		vals = append(vals, getValue(i))
-//		keys = append(keys, common.BigToHash(keyB).Bytes())
-//		keyB = keyB.Add(keyB, keyDelta)
-//		keyDelta.Add(keyDelta, common.Big1)
-//	}
-//	for i, k := range keys {
-//		nt.Update(k, common.CopyBytes(vals[i]))
-//	}
-//
-//	for i, k := range keys {
-//		blob, err := st.MarshalBinary()
-//		if err != nil {
-//			t.Fatal(err)
-//		}
-//		newSt, err := NewFromBinary(blob, nil)
-//		if err != nil {
-//			t.Fatal(err)
-//		}
-//		st = newSt
-//		st.Update(k, common.CopyBytes(vals[i]))
-//	}
-//	if have, want := st.Hash(), nt.Hash(); have != want {
-//		t.Fatalf("have %#x want %#x", have, want)
-//	}
-//}
+func TestStacktrieSerialization(t *testing.T) {
+	var (
+		st       = NewStackTrieWithOwner(nil, common.Hash{0x12})
+		nt       = NewEmpty(NewDatabase(rawdb.NewMemoryDatabase(), nil))
+		keyB     = big.NewInt(1)
+		keyDelta = big.NewInt(1)
+		vals     [][]byte
+		keys     [][]byte
+	)
+	getValue := func(i int) []byte {
+		if i%2 == 0 { // large
+			return crypto.Keccak256(big.NewInt(int64(i)).Bytes())
+		} else { //small
+			return big.NewInt(int64(i)).Bytes()
+		}
+	}
+	for i := 0; i < 10; i++ {
+		vals = append(vals, getValue(i))
+		keys = append(keys, common.BigToHash(keyB).Bytes())
+		keyB = keyB.Add(keyB, keyDelta)
+		keyDelta.Add(keyDelta, common.Big1)
+	}
+	for i, k := range keys {
+		nt.Update(k, common.CopyBytes(vals[i]))
+	}
+
+	for i, k := range keys {
+		blob, err := st.MarshalBinary()
+		if err != nil {
+			t.Fatal(err)
+		}
+		newSt, err := NewFromBinaryV2(blob, nil)
+		if err != nil {
+			t.Fatal(err)
+		}
+		st = newSt
+		st.Update(k, common.CopyBytes(vals[i]))
+	}
+	if have, want := st.Hash(), nt.Hash(); have != want {
+		t.Fatalf("have %#x want %#x", have, want)
+	}
+	if have, want := st.owner, (common.Hash{0x12}); have != want {
+		t.Fatalf("have %#x want %#x", have, want)
+	}
+}