This change refactors stacktrie to separate the stacktrie itself from the internal representation of nodes: a stacktrie is not a recursive structure of stacktries, rather, a framework for representing and operating upon a set of nodes.

I think it makes it easier to reason about, and also easier to deal with shared assets such as a hasher, the 'owner', and 'writeFunc' by having an envelope, rather than passing these things along at every constructor. 

This way of doing it also opens up for some new optimizations: e.g. we could do without global pool for nodes, and use a local non-threadsafe pool which is only accessible internally. (Example: https://github.com/ethereum/go-ethereum/commit/841824885fb6a68eb1b1b9f1449420d11a923599) 

~~This PR is not yet complete, and it also does away with the binary serialization. Not sure if it's something we want to keep: unused code has a tendency to bitrot.~~ The binary serialization has been added back, in it's own file. However, since the 'owner' of the trie now is moved to top level and not repeated in every node, the format was changed. Thus, I removed the original constructor so that this format-change must cause a compiler failure. But in general, I don't know if this really is something we want to kee, since we're not using it ourselves. 

We don't have an awful lot of benchmarks for the stacktrie, afaict only one really, which indicates a slight improvement in this PR (as for why the mem went up, that's curious, I can probably get that down again). 

```
name                       old time/op    new time/op    delta
DeriveSha200/std_trie-8       834µs ± 3%     848µs ± 1%    ~     (p=0.421 n=5+5)
DeriveSha200/stack_trie-8     514µs ± 5%     483µs ± 5%  -5.98%  (p=0.032 n=5+5)

name                       old alloc/op   new alloc/op   delta
DeriveSha200/std_trie-8       289kB ± 0%     290kB ± 0%  +0.09%  (p=0.008 n=5+5)
DeriveSha200/stack_trie-8    43.5kB ± 0%    46.3kB ± 0%  +6.48%  (p=0.008 n=5+5)

name                       old allocs/op  new allocs/op  delta
DeriveSha200/std_trie-8       2.91k ± 0%     2.92k ± 0%  +0.03%  (p=0.016 n=5+4)
DeriveSha200/stack_trie-8     1.25k ± 0%     1.26k ± 0%  +0.96%  (p=0.008 n=5+5)
```