@holiman 2023-10-09
Originally I think the binary marshalling was implemented so that we could pause/resume during sync. We didn't make use of that. @karalabe do you have any preference, whether to keep the feature or remove it? 

I am a bit loathe to keep maintaining behaviour that we don't use (and have no reason to believe that anyone else is using either)